pyproject.toml
[project]
name = "talkcouch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "gradio>=6.3.0",
    "langchain>=1.2.3",
    "langchain-google-genai>=4.1.3",
    "python-dotenv>=1.2.1",
    "speechrecognition>=3.14.5",
    "uvicorn>=0.40.0",
]
